# docs/openapi.yml
openapi: "3.0.3"
info:
  title: bloop API
  description: |
    Self-hosted error observability and LLM tracing API.

    bloop ingests error events from mobile apps and APIs, deduplicates them via
    fingerprinting, and provides query endpoints for reviewing errors. Optional
    feature-gated modules add DuckDB-powered analytics and LLM call tracing.

    **Conventions**

    - Timestamps are Unix epoch **milliseconds** unless noted otherwise.
    - LLM costs are stored as **microdollars** (1 dollar = 1,000,000 micros).
    - All error responses use `{ "error": "message" }`.
    - Pagination defaults: `limit=50`, `offset=0`, max `limit=200`.
  version: "1.0.0"
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development

# ---------------------------------------------------------------------------
# Security schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    hmac:
      type: apiKey
      in: header
      name: X-Bloop-Signature
      description: |
        HMAC-SHA256 signature computed over the raw request body using the
        project API key. Used by ingest endpoints.

    bearer:
      type: http
      scheme: bearer
      description: |
        Bearer token issued via the management API. Each token is scoped to a
        project and carries permission scopes such as `errors:read` and
        `errors:write`.

    session:
      type: apiKey
      in: cookie
      name: bloop_session
      description: |
        Cookie-based session set after WebAuthn passkey authentication.
        Used by dashboard and management endpoints.

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "validation failed: message is required"

    # -- Ingest --
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        api_key:
          type: string
          description: Only returned on create and key rotation.
        created_at:
          type: integer
          format: int64
        created_by:
          type: string
          nullable: true
    ProjectSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        created_at:
          type: integer
          format: int64
        created_by:
          type: string
          nullable: true
    CreateProject:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
          description: Lowercase alphanumeric with hyphens.
          pattern: "^[a-z0-9-]+$"
    UpdateProject:
      type: object
      properties:
        name:
          type: string

    # -- Alerts --
...
paths:
...
  /v1/projects:
    get:
      operationId: listProjects
      tags: [Projects]
      summary: List all projects
      description: Returns all projects without API keys.
      security:
        - session: []
      responses:
        "200":
          description: List of projects.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
    post:
      operationId: createProject
      tags: [Projects]
      summary: Create a new project
      description: Creates a project and generates an API key. The API key is only returned in this response.
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProject"
      responses:
        "200":
          description: Created project with API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error (empty name/slug, duplicate slug).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/projects/{slug}:
    get:
      operationId: getProject
      tags: [Projects]
      summary: Get project details
      security:
        - session: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details (without API key).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
        "404":
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProject
      tags: [Projects]
      summary: Update project
      security:
        - session: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProject"
      responses:
        "200":
          description: Updated project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
        "404":
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [Projects]
      summary: Delete a project
      description: Cannot delete the default project.
      security:
        - session: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deletion confirmed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                    example: true
        "400":
          description: Cannot delete the default project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...